fix: robustness sweep from whole-project review
- GS deferred acceptance: quota-blocked preferences are skipped and
  revisited, not permanently rejected; displacement can free quota
- Deterministic passes: applicants swept in id order, IA settles
  positions in id order (BTree collections, Ord on id newtypes)
- Event seq widened to u32 (u16 overflow panicked debug / wrapped release)
- CapacityStore::from_pool honors appeals: committed appealed seats stay
  quota-exempt across sync cycles
- Commands serialized behind one tokio Mutex held across DB I/O
  (add/remove_appeal TOCTOU with sync eliminated; no poisoning;
  matching runs via block_in_place)
- commit reloads the corpus and returns a receipt {inserted, snapshot};
  purge also clears cca_appeals and returns {deleted, snapshot};
  the review stepper survives the run being consumed
- Stale cca_appeals rows surface as snapshot warnings (toasted on sync)
- CLI prints an allocation summary again; dead appeals CSV dir removed

+++ crates/cupid-core/src/models/ledger.rs
@@ -9,7 +9,7 @@ use super::position::{Position, PositionIdx};
 pub struct Step {
     pub algorithm: Algorithm,
     pub round: u8,
-    pub seq: u16,
+    pub seq: u32,
 }
 
 /// The write surface used DURING a matching run.
